PR feedback: unit test conventions
