+++ com.unity.robotics.urdf-importer/Tests/Runtime/UrdfGeometryCollisionTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 870d199570ff640c78d5ada5c411ceed
+guid: e43314c00f437435ebc80f27ba220093
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
